Fix issue with scaling of input samples to float on encoder side. Use 48 kHz as default. Currently not supported for other Fs. Adding simulation data for plotting. Increment buffer size to max of 16 MB for encoded file.

+++ MPEG_Decoder/inc/init.h
@@ -59,7 +59,7 @@ extern uint8_t *pFRAME1;        // pointer reference
 extern short FRAME1[448];        // Rcv Frame
 extern short *pFRAME1;            // pointer reference
 #endif
-extern int32_t BSPL_rx[BANDSIZE];    // received bit values for subbands
+extern uint8_t BSPL_rx[BANDSIZE];    // received bit values for subbands
 extern float scf_rx[BANDSIZE];        // received bit values for scalefactors
 extern short tot_bits_rx;            // number of received bits
 extern short cnt_FRAME_read;        // array index for received data
